feat: Update project dependencies and configurations
This commit updates project dependencies to newer versions and adjusts build configurations for Vite and TypeScript. It also renames the project to "coffee-maroc" and modifies the `index.html` to include necessary script imports and a build output. The `vite.config.ts` is simplified, and TypeScript configurations are split into `tsconfig.app.json` and `tsconfig.node.json`.